requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
curl_cffi==0.12.0
cycler==0.12.1
discord.py==2.5.2
docopt==0.6.2
dotenv==0.9.9
et_xmlfile==2.0.0
Flask==3.1.1
fonttools==4.59.0
frozendict==2.4.6
frozenlist==1.7.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
multidict==6.6.3
multitasking==0.0.12
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pathlib==1.0.1
peewee==3.18.2
pillow==11.3.0
pipreqs==0.4.13
platformdirs==4.3.8
propcache==0.3.2
protobuf==6.31.1
psycopg2==2.9.10
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
six==1.17.0
soupsieve==2.7
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
yarg==0.1.10
yarl==1.20.1
yfinance==0.2.65
